[Refactor/#36] 중복 타입 제거
