+++ src/pages/NoteView.tsx
@@ -1,4 +1,6 @@
-import { useNoteView } from '@/hooks/useNoteView';
+import { notes } from '@/data/noteData';
+import { Note } from '@/store/noteStore';
+import { useState } from 'react';
 import styled from 'styled-components';
 import TagList from '@/components/note/TagList';
 import NoteList from '@/components/note/NoteList';
@@ -6,21 +8,51 @@ import Pagination from '@/components/note/Pagination';
 import NoteDetail from '@/components/note/NoteDetail';
 import BackButton from '@/components/common/BackButton';
 
+const sortedNotes: Note[] = [...notes].sort(
+  (a, b) => b.date.getTime() - a.date.getTime(),
+);
+
+const allTags = Array.from(new Set(sortedNotes.flatMap((note) => note.tags)));
+
 function NoteView() {
-  const {
-    selectedNote,
-    showAllTags,
-    selectedTags,
-    allTags,
-    currentNotes,
-    currentPage,
-    totalPages,
-    handleNoteClick,
-    handleToggleTags,
-    handleTagClick,
-    handleResetTags,
-    setCurrentPage,
-  } = useNoteView();
+  const [selectedNote, setSelectedNote] = useState<Note | null>(sortedNotes[0]);
+  const [showAllTags, setShowAllTags] = useState(false);
+  const [selectedTags, setSelectedTags] = useState<string[]>([]);
+  const [currentPage, setCurrentPage] = useState(1);
+  const itemsPerPage = 5;
+
+  const handleNoteClick = (note: Note) => {
+    setSelectedNote(note);
+  };
+
+  const handleToggleTags = () => {
+    setShowAllTags(!showAllTags);
+  };
+
+  const handleTagClick = (tag: string) => {
+    setSelectedTags((prevTags) =>
+      prevTags.includes(tag)
+        ? prevTags.filter((t) => t !== tag)
+        : [...prevTags, tag],
+    );
+    setCurrentPage(1);
+  };
+
+  const handleResetTags = () => {
+    setSelectedTags([]);
+    setCurrentPage(1);
+  };
+
+  const filteredNotes = selectedTags.length
+    ? sortedNotes.filter((note) =>
+        selectedTags.some((tag) => note.tags.includes(tag)),
+      )
+    : sortedNotes;
+
+  const indexOfLastItem = currentPage * itemsPerPage;
+  const indexOfFirstItem = indexOfLastItem - itemsPerPage;
+  const currentNotes = filteredNotes.slice(indexOfFirstItem, indexOfLastItem);
+  const totalPages = Math.ceil(filteredNotes.length / itemsPerPage);
 
   return (
     <NoteViewStyle showAllTags={showAllTags}>
